+++ R/TheilSen.R
@@ -183,8 +183,6 @@ TheilSen <- function(
   dec.place = 2,
   lab.frac = 0.99,
   lab.cex = 0.8,
-  x.relation = "same",
-  y.relation = "same",
   ref.x = NULL,
   ref.y = NULL,
   data.col = "cornflowerblue",
@@ -197,7 +195,6 @@ TheilSen <- function(
   slope.percent = FALSE,
   date.breaks = 7,
   date.format = NULL,
-  strip.position = "top",
   plot = TRUE,
   silent = FALSE,
   ...
@@ -311,12 +308,9 @@ TheilSen <- function(
     data.col = data.col,
     trend = trend,
     extra.args = extra.args,
-    x.relation = x.relation,
-    y.relation = y.relation,
     ref.x = ref.x,
     ref.y = ref.y,
     auto.text = auto.text,
-    strip.position = strip.position,
     date.breaks = date.breaks,
     date.format = date.format
   )
@@ -529,12 +523,9 @@ build_theilsen_plot <- function(
   data.col,
   trend,
   extra.args,
-  x.relation,
-  y.relation,
   ref.x,
   ref.y,
   auto.text,
-  strip.position,
   date.breaks,
   date.format
 ) {
@@ -638,10 +629,8 @@ build_theilsen_plot <- function(
     get_facet(
       type,
       extra.args,
-      scales = relation_to_facet_scales(x.relation, y.relation),
       auto.text = auto.text,
       drop = FALSE,
-      strip.position = strip.position,
       wd.res = extra.args$wd.res %||% 8
     ) +
     x_scale_fun(